Link API static data

+++ jolocom/container_bin/start.sh
@@ -5,6 +5,10 @@ cd "$(dirname "$(readlink -f "$0")")"
 cd ../demo-sso
 mkdir -p dist/img
 cd dist/img
-ln -s ../assets/js/app.bundle.js .
+if [ ! -e app.bundle.js ]; then
+    ln -s ../assets/js/app.bundle.js .
+    ln -s ../../claim_stepsWalked.png .
+    ln -s ../../stepsWalked.json .
+fi
 cd -
 yarn start
